% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NHANES4}
\alias{NHANES4}
\title{NHANES4 Dataset}
\format{
A data frame
}
\source{
National Health and Nutrition Examination Survey
}
\usage{
NHANES4
}
\description{
A dataset containing health biomarkers and survival data
}
\keyword{datasets}
