% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_disco.R
\name{cal_disco}
\alias{cal_disco}
\title{cal_disco}
\usage{
cal_disco(d4, var, ref, parallel = FALSE, cpp = FALSE, ncores = 4)
}
\arguments{
\item{d4}{Dataframe, containing age and biomarker information}

\item{var}{Enrolled biomarkers variables for analysis}

\item{ref}{Dataframe, young population information as a reference}

\item{parallel}{bool, use multiple cores for computation (default: FALSE)}

\item{cpp}{bool, use Rcpp with multiple cores for computation (default: FALSE)}

\item{ncores}{numeric, cpu cores for computation (default: 4)}
}
\value{
a data.frame with DISCO values
\itemize{
\item column DISCO, distance of covariance
}
}
\description{
Calculate distance of covariance for quantifying homeostatic dysregulations
}
\examples{
\dontrun{
data(NHANES4)
var <- c("age", "albumin", "alp", "creat", "glucose_mmol", "lymph", "mcv", "rdw", "wbc", "ggt")
NHANES4y <- NHANES4[which(NHANES4$age <= 30), ]
NHANES4m <- cal_disco(NHANES4, var, NHANES4y)
}
}
